Refactoring commit
Refactoring name and path for FeaturesOrganism to FeaturesMolecule
